[Android/Feat] 루틴 등록 기능 구현 - 전체 디바이스 정보를 통한 루틴 저장 S12P31D103-213

See merge request s12-final/S12P31D103!152

+++ Android/app/src/main/java/com/hogumiwarts/lumos/MainScreen.kt
@@ -24,7 +24,7 @@ import com.hogumiwarts.lumos.ui.navigation.BottomNavigation
 import com.hogumiwarts.lumos.ui.navigation.NavGraph
 
 @Composable
-fun MainScreen(deviceId: Long, deviceType: String) {
+fun MainScreen(deviceId: Int, deviceType: String) {
     val navController = rememberNavController()
 
     // 현재 백스택 엔트리 가져오기
