+++ Android/domain/src/main/java/com/hogumiwarts/domain/model/minibig/SwitchStatusData.kt
@@ -2,7 +2,7 @@ package com.hogumiwarts.domain.model.minibig
 
 data class SwitchStatusData(
     val tagNumber: Long,
-    val deviceId: Long,
+    val deviceId: Int,
     val deviceImg: String?,
     val deviceName: String,
     val manufacturerCode: String?,
